--- a/mp/src/game/client/c_baseplayer.cpp
+++ b/mp/src/game/client/c_baseplayer.cpp
@@ -493,7 +493,7 @@ void C_BasePlayer::Spawn( void )
 //-----------------------------------------------------------------------------
 // Purpose: 
 //-----------------------------------------------------------------------------
-bool C_BasePlayer::AudioStateIsUnderwater( Vector vecMainViewOrigin )
+bool C_BasePlayer::AudioStateIsUnderwater( const Vector &vecMainViewOrigin )
 {
 	if ( IsObserver() )
 	{