--- a/mp/src/game/server/baseentity.cpp
+++ b/mp/src/game/server/baseentity.cpp
@@ -3715,8 +3715,9 @@ void CBaseEntity::ComputeWorldSpaceSurroundingBox( Vector *pMins, Vector *pMaxs 
 //------------------------------------------------------------------------------
 const char *CBaseEntity::GetDebugName(void)
 {
-	if ( this == NULL )
-		return "<<null>>";
+    // if (!this) is always false
+	/*if ( this == NULL )
+		return "<<null>>";*/
 
 	if ( m_iName != NULL_STRING ) 
 	{