--- a/mp/src/game/server/momentum/tickset.h
+++ b/mp/src/game/server/momentum/tickset.h
@@ -29,10 +29,10 @@ public:
             this->sType = other.sType;
             return *this;
         }
-        bool operator ==(const Tickrate &other)
+        bool operator ==(const Tickrate &other) const
         {
-            return (mom_UTIL->FloatEquals(other.fTickRate, fTickRate)
-                && !Q_strcmp(other.sType, sType));
+            return mom_UTIL->FloatEquals(other.fTickRate, fTickRate)
+                && !Q_strcmp(other.sType, sType);
         }
     };