… mp/src/game/server/momentum/mom_replay_entity.cpp
… mp/src/game/server/momentum/mom_triggers.cpp
… mp/src/game/server/momentum/mom_triggers.h
… mp/src/game/server/player.h
… mp/src/game/server/props.cpp
… mp/src/game/server/subs.cpp
… mp/src/game/server/trains.cpp
… mp/src/game/server/util.cpp
… mp/src/game/server/util.h
… mp/src/game/shared/basegrenade_shared.cpp
… mp/src/game/shared/collisionproperty.cpp
… mp/src/game/shared/mapentities_shared.cpp
… mp/src/game/shared/momentum/mom_gamemovement.cpp
… mp/src/game/shared/momentum/mom_shareddefs.h
… mp/src/game/shared/momentum/util/baseautocompletefilelist.cpp
… mp/src/game/shared/momentum/util/mom_util.cpp
… mp/src/game/shared/momentum/util/mom_util.h
… mp/src/game/shared/momentum/weapon/weapon_basecsgrenade.cpp
… mp/src/game/shared/momentum/weapon/weapon_basecsgrenade.h
… mp/src/game/shared/momentum/weapon/weapon_csbase.cpp
… mp/src/game/shared/momentum/weapon/weapon_knife.cpp
… mp/src/game/shared/momentum/weapon/weapon_mom_grenade.cpp
… mp/src/game/shared/momentum/weapon/weapon_mom_grenade.h
… mp/src/game/shared/props_shared.cpp
… mp/src/game/shared/weapon_parse.cpp
… mp/src/public/collisionutils.cpp
… mp/src/public/raytrace.h
… mp/src/public/studio.cpp
… mp/src/public/tier1/memstack.h
… mp/src/public/tier1/utlblockmemory.h