Branched form the bugfix branch, the code was inspected with PVS-Studio and some errors were fixed, some optimizations done, and few improvements coded.

Not really exciting, but for Counter-Strike:Source's test_hardware.bsp map, the fps numbers look like:

|STAGE|MIN|AVG|MAX|
|-|-|-|-|   
|PRE OPTIMIZATION|44 fps|959.7 fps|1872 fps|  
|POST OPTIMIZATION|48 fps|965.6 fps|1876 fps| 

There were some... interesting fuckups from VALVe here and there.

We could do another one once we're near Beta. There were **thousands** of issues not resolved in this one, as I only focused on what seemed more important.